[Setup] Remove toast notifications and other dependencies from bootstrapper
