+++ PowerToys.sln
@@ -281,8 +281,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "utils", "utils", "{B39DC643
 		src\common\utils\EventLocker.h = src\common\utils\EventLocker.h
 		src\common\utils\EventWaiter.h = src\common\utils\EventWaiter.h
 		src\common\utils\exec.h = src\common\utils\exec.h
+		src\common\utils\HttpClient.h = src\common\utils\HttpClient.h
 		src\common\utils\json.h = src\common\utils\json.h
 		src\common\utils\logger_helper.h = src\common\utils\logger_helper.h
+		src\common\utils\MsiUtils.h = src\common\utils\MsiUtils.h
 		src\common\utils\os-detect.h = src\common\utils\os-detect.h
 		src\common\utils\process_path.h = src\common\utils\process_path.h
 		src\common\utils\ProcessWaiter.h = src\common\utils\ProcessWaiter.h
