+++ installer/PowerToysBootstrapper/bootstrapper/DotnetInstallation.h
