+++ installer/PowerToysBootstrapper/bootstrapper/pch.h
@@ -6,6 +6,7 @@
 #include <shellapi.h>
 #include <commctrl.h>
 
+#include <charconv>
 #include <string_view>
 #include <optional>
 #include <fstream>
@@ -22,8 +23,10 @@
 #include <spdlog/sinks/null_sink.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
 
+#pragma warning(push, 0)
 #include <winrt/base.h>
 #include <winrt/Windows.Foundation.h>
 #include <winrt/Windows.Foundation.Collections.h>
+#pragma warning(pop)
 
 #include <cxxopts.hpp>
